# specs/001-widget-auth-admin-cicd/contracts/widget-admin.openapi.yaml
openapi: 3.1.0
info:
  title: Albert — Tenant Admin Widget Configuration
  version: 0.1.0
  description: |
    Tenant-admin endpoints consumed by the Streamlit admin app.

    Authentication: Bearer JWT from POST /api/v1/auth/login. The user must
    hold a tenant_membership.role of "tenant_admin" for the tenant in path
    or query — never inferred from the request body.

    Tenant scope is set on the request from the user's resolved membership;
    cross-tenant access (e.g. user from Tenant A asks about Tenant B's
    widget id) MUST return 404, not 403 (do not confirm existence).
paths:
  /api/v1/admin/widgets:
    get:
      operationId: listWidgets
      summary: List widgets owned by the calling user's tenant.
      security:
        - platformAuth: []
      responses:
        '200':
          description: Widgets owned by the caller's tenant.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdminWidget' }
    post:
      operationId: createWidget
      summary: Create a widget for the caller's tenant.
      security:
        - platformAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWidgetRequest'
      responses:
        '201':
          description: Widget created. Server-generated public_widget_id is returned.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminWidget' }
  /api/v1/admin/widgets/{widget_id}:
    parameters:
      - in: path
        name: widget_id
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updateWidget
      summary: Edit theme / greeting / status.
      security:
        - platformAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateWidgetRequest' }
      responses:
        '200': { description: Updated. }
        '404': { description: Not found OR not owned by caller's tenant. }
...
components:
  securitySchemes:
    platformAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AdminWidget:
      type: object
      required: [id, public_widget_id, name, theme, greeting, status]
      properties:
        id: { type: string, format: uuid }
        public_widget_id: { type: string }
        name: { type: string }
        theme: { type: object, additionalProperties: true }
        greeting: { type: string }
        status: { type: string, enum: [enabled, disabled] }
    CreateWidgetRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        greeting: { type: string, maxLength: 500 }
        theme: { type: object, additionalProperties: true }
    UpdateWidgetRequest:
      type: object
      additionalProperties: false
      properties:
        name: { type: string }
        greeting: { type: string, maxLength: 500 }
        theme: { type: object, additionalProperties: true }
        status: { type: string, enum: [enabled, disabled] }
...